added deamon skeleton
