+++ node-net-manager/go.mod
@@ -7,8 +7,10 @@ require (
 	github.com/eclipse/paho.mqtt.golang v1.4.3
 	github.com/google/gopacket v1.1.19
 	github.com/gorilla/mux v1.8.1
+	github.com/kardianos/service v1.2.2
 	github.com/sipcapture/heplify v1.65.16
 	github.com/songgao/water v0.0.0-20200317203138-2b4b6d7c09d8
+	github.com/spf13/cobra v1.8.1
 	github.com/tkanos/gonfig v0.0.0-20210106201359-53e13348de2f
 	github.com/vishvananda/netlink v1.2.1-beta.2
 	github.com/vishvananda/netns v0.0.4
@@ -19,8 +21,10 @@ require (
 	github.com/ghodss/yaml v1.0.0 // indirect
 	github.com/google/go-cmp v0.6.0 // indirect
 	github.com/gorilla/websocket v1.5.1 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/rogpeppe/go-internal v1.11.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/stretchr/testify v1.8.4 // indirect
 	golang.org/x/net v0.22.0 // indirect
 	golang.org/x/sync v0.6.0 // indirect
